+++ apps/picsa-tools/extension-content/src/app/pages/home/extension-home.component.ts
@@ -5,7 +5,6 @@ import { AfterViewInit, Component, ElementRef, OnDestroy, ViewChild } from '@ang
 import { MatIconModule } from '@angular/material/icon';
 import { Router, RouterModule } from '@angular/router';
 import { marker as translateMarker } from '@biesbjerg/ngx-translate-extract-marker';
-import { AppUpdate, AppUpdateAvailability, AppUpdateInfo } from '@capawesome/capacitor-app-update';
 import { IonicModule } from '@ionic/angular';
 import { PicsaCommonComponentsModule, PicsaCommonComponentsService } from '@picsa/components';
 import { ConfigurationService } from '@picsa/configuration/src';
@@ -150,8 +149,6 @@ export class ExtensionHomeComponent implements AfterViewInit, OnDestroy {
   public additionalLinks = ADDITIONAL_LINKS;
   public version = APP_VERSION;
 
-  public appUpdateInfo: AppUpdateInfo | undefined;
-
   @ViewChild('headerContent')
   headerContent: ElementRef<HTMLElement>;
 
@@ -179,29 +176,6 @@ export class ExtensionHomeComponent implements AfterViewInit, OnDestroy {
     this.componentsService.patchHeader({
       cdkPortalEnd: new DomPortal(this.headerContent),
     });
-
-    // Check for updates using Google Play API
-    AppUpdate.getAppUpdateInfo().then((appUpdateInfo: AppUpdateInfo) => {
-      if (
-        appUpdateInfo.updateAvailability === AppUpdateAvailability.UPDATE_AVAILABLE &&
-        appUpdateInfo.flexibleUpdateAllowed
-      ) {
-        this.startFlexibleUpdate();
-      }
-    });
-  }
-
-  private async startFlexibleUpdate() {
-    try {
-      await AppUpdate.startFlexibleUpdate();
-      console.log('Flexible update started');
-
-      // Complete the update once it's downloaded
-      await AppUpdate.completeFlexibleUpdate();
-      console.log('Flexible update completed successfully');
-    } catch (error) {
-      console.error('Error during flexible update:', error);
-    }
   }
 
   public startTour() {
