+++ libs/shared/src/services/native/index.ts
@@ -1,4 +1,5 @@
 // NOTE - to use these service must import PicsaNativeModule in app root with forRoot call
 
+export * from './app-update';
 export * from './print';
 export * from './storage-service';
